Add tests
Right now I am just testing the Telegram accepts the message with the
specified parse modes.

MarkdownV2 mode is particularly picky and you can easily get a 400
status code (bad request) if you use markdown with a bad syntax or
an unsupported feature.

refs: zmoog/til#15
