+++ tests/cassettes/test_message_send/test_send_html.yaml
@@ -11,12 +11,12 @@ interactions:
       User-Agent:
       - python-requests/2.28.2
     method: GET
-    uri: https://api.telegram.org/1234567890abcdef1234567890abcdef/sendMessage?chat_id=123456&text=%27Hello%2C+world%21%27
+    uri: https://api.telegram.org/1234567890abcdef1234567890abcdef/sendMessage?chat_id=123456&text=welcome+to+this+%3Cb%3Efucking%3C%2Fb%3E+world&parse_mode=HTML
   response:
     body:
-      string: '{"ok":true,"result":{"message_id":665,"from":{"id":123,"is_bot":true,"first_name":"Hogwarts
-        Bot","username":"hogwarts_api_bot"},"chat":{"id":456,"first_name":"Maurizio","last_name":"Branca","username":"zmoog","type":"private"},"date":1675921959,"text":"''Hello,
-        world!''"}}'
+      string: '{"ok":true,"result":{"message_id":721,"from":{"id":1051994739,"is_bot":true,"first_name":"Hogwarts
+        Bot","username":"hogwarts_api_bot"},"chat":{"id":123456,"first_name":"Maurizio","last_name":"Branca","username":"zmoog","type":"private"},"date":1676700426,"text":"welcome
+        to this fucking world","entities":[{"offset":16,"length":7,"type":"bold"}]}}'
     headers:
       Access-Control-Allow-Methods:
       - GET, POST, OPTIONS
@@ -27,11 +27,11 @@ interactions:
       Connection:
       - keep-alive
       Content-Length:
-      - '284'
+      - '350'
       Content-Type:
       - application/json
       Date:
-      - Thu, 09 Feb 2023 05:52:39 GMT
+      - Sat, 18 Feb 2023 06:07:06 GMT
       Server:
       - nginx/1.18.0
       Strict-Transport-Security:
